backend/requirements-dev.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
sqlalchemy==2.0.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.28.2
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.0
pytest==7.3.1
pytest-asyncio==0.21.0
